IRIS-3881 | removed AlmostTherePageObject, ConfirmationPageObject and refactored tests to use new email confirmation flow that doesn't require additional credentials providing
