+++ src/test/java/com/wikia/webdriver/pageobjectsfactory/pageobject/UserProfilePageObject.java
@@ -1,4 +1,4 @@
-package com.wikia.webdriver.pageobjectsfactory.pageobject.signup;
+package com.wikia.webdriver.pageobjectsfactory.pageobject;
 
 import com.wikia.webdriver.common.contentpatterns.URLsContent;
 import com.wikia.webdriver.common.core.AlertHandler;
